+++ examples/petstore/types.rb
@@ -4,16 +4,16 @@
 require "keiyaku/runtime"
 
 module Petstore
-  Order = Keiyaku.model(
+  Order = Keiyaku.model({
     id: Integer,
     pet_id: Integer,
     quantity: Integer,
     ship_date: Time,
     status: String,
     complete: :bool
-  )
-  Category = Keiyaku.model(id: Integer, name: String)
-  User = Keiyaku.model(
+  })
+  Category = Keiyaku.model({ id: Integer, name: String })
+  User = Keiyaku.model({
     id: Integer,
     username: String,
     first_name: String,
@@ -22,16 +22,15 @@ module Petstore
     password: String,
     phone: String,
     user_status: Integer
-  )
-  Tag = Keiyaku.model(id: Integer, name: String)
-  Pet = Keiyaku.model(
+  })
+  Tag = Keiyaku.model({ id: Integer, name: String })
+  Pet = Keiyaku.model({
     id: Integer,
     name: String,
     category: Category,
     photo_urls: [String],
     tags: [Tag],
-    status: String,
-    required: %i[name photo_urls]
-  )
-  ApiResponse = Keiyaku.model(code: Integer, type: String, message: String)
+    status: String
+  }, required: %i[name photo_urls])
+  ApiResponse = Keiyaku.model({ code: Integer, type: String, message: String })
 end